+++ src/pages/WeatherDisplay.js
@@ -10,6 +10,11 @@ import {
 } from 'react-icons/fa';
 import { Spin } from 'antd';
 import { LoadingOutlined } from '@ant-design/icons';
+import moment from 'moment';
+import 'moment/locale/zh-cn';
+
+moment.locale('zh-cn');
+
 const WeatherDisplay = ({ weatherText, futureWeatherText }) => {
   const getWeatherIcon = (weather) => {
     switch (weather) {
@@ -89,30 +94,15 @@ const WeatherDisplay = ({ weatherText, futureWeatherText }) => {
           const futureDate = new Date(today);
           futureDate.setDate(today.getDate() + index + 1);
 
-          let dateDescription;
-          switch (index + 1) {
-            case 1:
-              dateDescription = '明天';
-              break;
-            case 2:
-              dateDescription = '后天';
-              break;
-            case 3:
-              dateDescription = '大后天';
-              break;
-            default:
-              dateDescription = `${
-                futureDate.getMonth() + 1
-              }/${futureDate.getDate()}`;
-              break;
-          }
           const dayWeatherIcon = getWeatherIcon(dayweather);
+
           return (
             <div key={index}>
-              <div>{dateDescription}</div>
+              <div>{index + 1 === 1 && '明天'}</div>
+              <div>{index + 1 === 2 && '后天'}</div>
+              <div>{index + 1 === 3 && '大后天'}</div>
 
               <div>{dayWeatherIcon}</div>
-
               <div>{dayweather}</div>
               <div>{`${nighttemp} - ${daytemp}°`}</div>
             </div>
