+++ src/pages/api/storage.js
@@ -29,40 +29,5 @@ export default async function handler(req, res) {
       console.error('Error executing query:', error);
       res.status(500).json({ error: 'Internal server error' });
     }
-  } else if (req.method === 'GET') {
-    try {
-      const client = await pool.connect();
-      const query = 'SELECT * FROM dreams';
-      const result = await client.query(query);
-
-      client.release();
-
-      res.status(200).json(result.rows);
-    } catch (error) {
-      console.error('Error executing query:', error);
-      res.status(500).json({ error: 'Internal server error' });
-    }
-  } else if (req.method === 'PUT') {
-    const { id, status } = req.body;
-
-    if (id !== undefined && status !== undefined) {
-      try {
-        const client = await pool.connect();
-        const query = 'UPDATE dreams SET status = $1 WHERE id = $2';
-        const values = [status, id];
-        await client.query(query, values);
-
-        client.release();
-
-        res.status(200).json({ message: 'Record updated' });
-      } catch (error) {
-        console.error('Error updating record:', error);
-        res.status(500).json({ error: 'Internal server error' });
-      }
-    } else {
-      res.status(400).json({ error: 'ID and status are required for update' });
-    }
-  } else {
-    res.status(405).json({ error: 'Method Not Allowed' });
   }
 }
